poll connection
A special connection that can be used to poll for the termination
of the process. It is created by default if neither stdout, nor
stderr are written to a pipe, and no extra connections are passed
to the child process. It can also be forced with the `poll_connection`
argument.

So $poll_io() and poll() now return three-element character vectors
for processes: 'input', 'output' and 'process'.

Closes #50, closes #101.