Fix extra connection bug on Unix
The extra connections that we pass to the child,
need to be closed with the processx connection API,
because otherwise they will be closed (again) when
their finalizer runs, and in the meanwhile the FD
might be allocated for something else.

Now we close the extra processx connections properly
in the parent.