Send interrupt to background process (#127)
* Send interrupt to process

Unix, #126

* Win: fix non-detached process cleanup

We need to attach it to the global job object
if it is not detached. This will clean it up
when the R process exits.

* Implement interrupt on Windows

Sends CTRL+C currently, but easily configurable
to send CTRL+BREAK, should we need that.

This is a bit difficult to implement on windows,
because you need to attach the console of the
target process, which means that the calling
process will lose its console. So we create a
subprocess just for this, this is apparently
the only clean solution. We run the child
process synchronously, it is fairly fast, about
~ 0.05s.

* Bump dev version number