Work around IOCP crash by marking the key
processx uses IOCPs to implement poll() on Windows.
One trouble with IOCPs is that on thread can only be
associated with one IOCP. If you create another IOCP
while the previous one is still active, the result
is apparently a mess.

R-devel uses IOCPs now,
https://github.com/wch/r-source/commit/fd4545643221862555559da9f296e3f996bfc567
which is a problem for processx.

The proper solution is to do all processx poll/read
in another thread. However this is a quick fix that
at least avoids some crashes. This fix simply ignores
async reads that were not started by processx.

It works by marking all processx connections with a
random integer as the first element of the struct.
After an async read completes, we check if it has the
marker, and if not we simply ignore it.

This fixes #184.