Fix cleanup bugs on Windows
When the handle is finalized at the end of the
session, there might still be references to it,
and if these references also have finalizer, they
might call our processx finalizer again, or any
other methods. So we need to be ready for this.

In particular, the async worker pool has a finalizer
that kills all worker processes, but by the time
this is called, the process finalizers have already
kill them.