Do not change inherited extra connections
Also, new `nonblocking` argument to `conn_create_pipepair()`.
We don't touch the inherited file descriptors any more (Unix),
because not all FDs can be set to non-blocking, and not all
FDs should be closed in the parent.

It is now the user's reponsibility to close the FDs, and also
to set the appropriate non-blocking or blocking flag. The latter
is simpler now for pipe pairs, with the new `nonblocking` argument
to `conn_create_pipepair()`.