Better printing of errors and traces
The `conditionMessage()` method now only contains
the message, but not the stderr. This is better, because
the trace prints the message, but does not duplicate
the stderr.

In general, it seems good practice to create a conditionMessage()
method that only returns the message, without lengthy output,
and also define the format & print methods to return/print all
information.