Fix boot- & create-time issues on Linux (#260)
Apparently, on Linux the boot time in /proc (btime in /proc/stat)
is not fixed, but it can change if the system clock changes.
In Docker it also changes if the host machine goes to sleep and
then wakes up. This is a problem for us, because we use
(pid, start-time) as a process id, and the start-time depends
on the boot time. ps caches the boot time, but processx queried
it every time it needed it, and then it possibly ended up with
a different create-time than ps, and the $as_ps_handle() method
created a ps handle that did not actually work, because ps thought
that the process has finished already.

So now we query the boot time via ps in .onLoad(), and use that
for the current session. This way the ps and processx boot times
always match.

Closes #258.