Make sure SIGCHLD is handled by main thread
Otherwise we can get all kinds of issues. The return status of
the process might be NA (as in #308), there might be double frees,
crashes, freezes. Only comes out if a SIGCHLD arrives while
we are blocking it, and there is another thread that is willing
to handle it. So it is rare, but certainly possible.

Closes #308.